+++ packages/api/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.20"></a>
+## [1.0.20](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.20-unstable.0...@aws-amplify/api@1.0.20) (2018-11-01)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/api
+
+<a name="1.0.20-unstable.0"></a>
+## [1.0.20-unstable.0](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.19...@aws-amplify/api@1.0.20-unstable.0) (2018-10-30)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/api
+
 <a name="1.0.19"></a>
 ## [1.0.19](https://github.com/aws/aws-amplify/compare/@aws-amplify/api@1.0.19-unstable.1...@aws-amplify/api@1.0.19) (2018-10-29)