+++ packages/storage/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.0.18"></a>
+## [1.0.18](https://github.com/aws/aws-amplify/compare/@aws-amplify/storage@1.0.18-unstable.0...@aws-amplify/storage@1.0.18) (2018-11-01)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/storage
+
+<a name="1.0.18-unstable.0"></a>
+## [1.0.18-unstable.0](https://github.com/aws/aws-amplify/compare/@aws-amplify/storage@1.0.17...@aws-amplify/storage@1.0.18-unstable.0) (2018-10-30)
+
+
+
+
+**Note:** Version bump only for package @aws-amplify/storage
+
 <a name="1.0.17"></a>
 ## [1.0.17](https://github.com/aws/aws-amplify/compare/@aws-amplify/storage@1.0.17-unstable.1...@aws-amplify/storage@1.0.17) (2018-10-29)